+++ src/aioway/_comps/iters.py
